requirements.txt
###### Requirements with Version Specifiers ######
# pkg_2 == 1.0.0
# pkg_3 != 1.0.0
# pkg_4 >= 1.0.0
# pkg_5 >= 1.0.0, <=2.0.0

selenium == 3.141.0
pyyaml
paramiko==2.8.1
requests == 2.26.0
setuptools-rust == 1.1.1
robotframework >= 4.1.2
robotframework-pythonlibcore == 3.0.0
robotframework-seleniumlibrary >= 5.1.3
robotframework-sshlibrary >= 3.7.0
robotframework-requests >= 0.9.2
